Added addressing modes and basic clock

+++ src/main/nes/Memory.java
@@ -4,12 +4,12 @@ import java.util.Arrays;
 
 public class Memory extends BusDevice {
 
-	int[] fakeRam = new int[64 * 1024];
+	byte[] fakeRam = new byte[64 * 1024];
 
 	public Memory() {
 		super(0x0000, 0xFFFF);
 
-		Arrays.fill(fakeRam, 0);
+		Arrays.fill(fakeRam, (byte) 0);
 	}
